refactor ui logic into hooks
