+++ site/src/app/ui.tsx
@@ -1,27 +1,10 @@
 'use client'
 
-import { calculateViewport } from '@/viewport'
-import * as Equivalence from '@effect/schema/Equivalence'
-import * as S from '@effect/schema/Schema'
-import { useSelector } from '@xstate/react'
-import * as Data from 'effect/Data'
-import * as Equal from 'effect/Equal'
-import { panels } from 'widgetui'
 import * as WidgetUI from 'widgetui/schema'
-import { createActor } from 'xstate'
-import { Element, UiContext, UiSnapshot, uiMachine } from './ui.machine'
-
-const ui = createActor(uiMachine, {
-  input: { panels },
-})
-ui.start()
-
-const selectIds = (state: UiSnapshot) => state.context.ids
-const compareIds = (a: string[], b: string[]) =>
-  Equal.equals(Data.array(a), Data.array(b))
+import { useElement, useIds, useViewport } from './ui.actor'
 
 export function UI() {
-  const ids = useSelector(ui, selectIds, compareIds)
+  const ids = useIds()
 
   return (
     <svg id="ui" className="debug" viewBox="0 0 3840 2160">
@@ -39,68 +22,6 @@ export function UI() {
 
 function Use({ id }: { id: string }) {
   return <use href={`#${id}`} />
-}
-
-const selectElement = (id: string) => (state: UiSnapshot) =>
-  state.context.elementById.get(id)!
-const compareElement = Equivalence.make(
-  S.union(WidgetUI.Collection, WidgetUI.Widget),
-)
-
-const useElement = (id: string): Element => {
-  const element = useSelector(ui, selectElement(id), compareElement)
-  return element
-}
-
-const selectParentByElement = (state: UiSnapshot) =>
-  state.context.parentByElement
-const compareParentByElement = (
-  a: UiContext['parentByElement'],
-  b: UiContext['parentByElement'],
-) => Equal.equals(Data.struct(a), Data.struct(b))
-
-const useViewport = (
-  element: Element,
-): { x: number; y: number; width: number; height: number } | undefined => {
-  const parentByElement = useSelector(
-    ui,
-    selectParentByElement,
-    compareParentByElement,
-  )
-
-  if (WidgetUI.isCollection(element)) {
-    return calculateViewport(element.ViewPort)
-  }
-
-  if (WidgetUI.isViewportWidget(element)) {
-    const viewports = [element.ViewPort]
-
-    let ancestor = parentByElement.get(element)
-    while (ancestor !== undefined) {
-      if (
-        WidgetUI.isCollection(ancestor) ||
-        WidgetUI.isViewportWidget(ancestor)
-      ) {
-        viewports.unshift(ancestor.ViewPort)
-      }
-      ancestor = parentByElement.get(ancestor)
-    }
-
-    const combinedViewport = {
-      ...calculateViewport(viewports.shift()!),
-      width: element.ViewPort.width,
-      height: element.ViewPort.height,
-    }
-    for (const viewport of viewports) {
-      const { x, y } = calculateViewport(viewport)
-      combinedViewport.x += x
-      combinedViewport.y += y
-    }
-
-    return combinedViewport
-  }
-
-  return undefined
 }
 
 function Def({ id }: { id: string }) {
