+++ site/src/app/ui.types.ts
@@ -1,0 +1,3 @@
+import { Collection, Widget } from 'widgetui/schema'
+
+export type Element = Collection | Widget
